backend/requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
pydantic==2.7.4
motor==3.5.0
python-dotenv==1.0.1
openai>=1.0.0